ui: move to namespace components to root of component directories
This commit moves the namespace components directory, because the idea of ​​the
project to keep the components in the root.
