+++ ui/src/components/namespace/NamespaceDelete.vue
