+++ ui/src/components/namespace/NamespaceInstructions.vue
@@ -69,7 +69,7 @@
 
 <script>
 
-import NamespaceAdd from '@/components/app_bar/namespace/NamespaceAdd';
+import NamespaceAdd from '@/components/namespace/NamespaceAdd';
 
 export default {
   name: 'NamespaceInstructionsComponent',
