+++ ui/src/components/namespace/NamespaceMemberDelete.vue
