+++ ui/src/components/namespace/NamespaceMemberFormDialog.vue
